requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alpaca==1.0.0
alpaca-py==0.39.0
# talib-binary==0.4.26
alpaca-trade-api==3.2.0
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backtesting==0.6.3
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
bokeh==3.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-SocketIO==5.5.1
Flask-Swagger-UI==4.11.1
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.33.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
lxml==5.3.2
MarkupSafe==3.0.2
#matplotlib==3.8.3
#matplotlib-inline==0.1.7
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
narwhals==1.31.0
nest-asyncio==1.6.0
nltk==3.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
sseclient-py==1.8.0
stack-data==0.6.3
tabulate==0.9.0
textblob==0.17.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.0
wordcloud==1.9.3
wsproto==1.2.0
xyzservices==2025.1.0
yarl==1.18.3
yfinance==0.2.54
python-dotenv==1.1.0
numpy==1.26.4
xgboost==3.0.0
scikit-learn==1.6.1
scipy==1.15.2
redis
